Update Clear Storage toast
