--- src/components/notifications/storage-full-notification.module.scss
@@ -1,36 +1,0 @@
-@import '../../styles/variables.module.scss';
-
-.buttonRow {
-  display: flex;
-  flex-direction: column-reverse;
-  justify-content: flex-end;
-  gap: 20px;
-  margin-top: 20px;
-
-  @media (min-width: $min-sm) {
-    flex-direction: row;
-  }
-}
-
-.cancelButton {
-  cursor: pointer;
-  border: none;
-  background: none;
-  font-weight: 400;
-}
-
-.clearButton {
-  cursor: pointer;
-  background-color: #194aff;
-  color: $primary-color;
-  min-width: 15ch;
-  padding: 12px 10px;
-  font-weight: 500;
-  border-radius: 4px;
-  border: none;
-  white-space: nowrap;
-
-  &:disabled {
-    opacity: 0.8;
-  }
-}
